fix(system-test-java): make money/rate assertions BigDecimal-exact
Settles BigDecimal as the canonical type for money/rate assertions in
system-test/java: String and double are thin delegating overloads, and
comparison is always isEqualByComparingTo.

- GetTaxVerification.taxRate: isEqualTo -> isEqualByComparingTo. This was
  the only scale-sensitive money comparison in the project; 0.10 failed
  against 0.1 with a message reading "Expected tax rate to be 0.1, but was
  0.1". The scale comes from JSON deserialization, not from the test, so
  nothing protected it.
- BrowseCouponsResponse.CouponDto.discountRate: double -> BigDecimal, and
  CouponManagementPage.parseDiscountRate now returns BigDecimal using
  movePointLeft(2) instead of double division by 100. This removes the last
  raw-double money comparison in system-test/java. Test-side driver DTO
  only; invisible outside system-test/java.
- BrowseCouponsVerification.couponHasDiscountRate: BigDecimal canonical
  plus a double overload. No String overload added -- no caller needs one.
- ViewOrderVerification.subtotalPrice(String): toDouble -> toBigDecimal.
- Converter.fromDouble: toString() -> toPlainString(), so setup steps never
  send scientific notation to the SUT.

Verified against gh-optivem-multitier-java.yaml: BrowseCouponsPositiveTest
(UI and API channels), PublishCoupon positive/negative, TaxStubContractTest,
TaxRealContractTest, and the PlaceOrder tax/coupon scenarios all pass. The
--sample sweep passed but did not cover these paths, so the targeted run is
the real evidence.

Plan 20260720-1420 executed and deleted; its remaining consistency item and
its findings are folded into 20260720-1350.

+++ system-test/java/src/main/java/com/mycompany/myshop/testkit/driver/port/dtos/BrowseCouponsResponse.java
@@ -5,6 +5,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 import java.util.List;
 
@@ -21,7 +22,7 @@ public class BrowseCouponsResponse {
     @AllArgsConstructor
     public static class CouponDto {
         private String code;
-        private double discountRate;
+        private BigDecimal discountRate;
         private Instant validFrom;
         private Instant validTo;
         private Integer usageLimit;
